# f3-server/crates/adapter_in_web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "adapter_in_web",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/after-login/upload": {
      "post": {
        "tags": [
          "crate::controller::file_upload_controller"
        ],
        "operationId": "upload_file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "アップロード成功"
          },
          "400": {
            "description": "アップロード失敗"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "crate::controller::auth_controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorCredential"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "ログイン成功"
          },
          "401": {
            "description": "認証情報が正しくない"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "crate::controller::auth_controller"
        ],
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "ログアウト成功（未認証状態で実行しても成功扱い）"
          }
        }
      }
    },
    "/health/db": {
      "get": {
        "tags": [
          "crate::controller::health_controller"
        ],
        "operationId": "health_check_db",
        "responses": {
          "200": {
            "description": "DBが正常に起動している"
          },
          "500": {
            "description": "DBが起動していない、またはクエリの実行に失敗した"
          }
        }
      }
    },
    "/listener/create": {
      "post": {
        "tags": [
          "crate::controller::listener_controller"
        ],
        "operationId": "create_listener",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListenerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Listenerの作成に成功"
          },
          "500": {
            "description": "Listenerの作成に失敗"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "Admin",
              "Write"
            ]
          }
        ]
      }
    },
    "/listener/list": {
      "get": {
        "tags": [
          "crate::controller::listener_controller"
        ],
        "operationId": "list_listeners",
        "responses": {
          "200": {
            "description": "Listenerの一覧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListListenerResponse"
                }
              }
            }
          },
          "500": {
            "description": "ハンドリングできない異常"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/listener/remove": {
      "post": {
        "tags": [
          "crate::controller::listener_controller"
        ],
        "operationId": "remove_listener",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveListenerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Listenerの削除に成功"
          },
          "500": {
            "description": "Listenerの削除に失敗"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "Admin",
              "Write"
            ]
          }
        ]
      }
    },
    "/listener/start": {
      "post": {
        "tags": [
          "crate::controller::listener_controller"
        ],
        "operationId": "start_listener",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartListenerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Listenerが正常に起動"
          },
          "500": {
            "description": "Listenerの起動に失敗"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "Admin",
              "Write"
            ]
          }
        ]
      }
    },
    "/listener/stop": {
      "post": {
        "tags": [
          "crate::controller::listener_controller"
        ],
        "operationId": "stop_listener",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopListenerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Listenerが正常に停止"
          },
          "500": {
            "description": "Listenerの停止に失敗"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "Admin",
              "Write"
            ]
          }
        ]
      }
    },
    "/operator/get": {
      "post": {
        "tags": [
          "crate::controller::operator_controller"
        ],
        "operationId": "get_operator",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetOperatorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "オペレータ詳細情報"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/operator/list": {
      "post": {
        "tags": [
          "crate::controller::operator_controller"
        ],
        "operationId": "list_operators",
        "responses": {
          "200": {
            "description": "オペレータ一覧"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/operator/toggle_status": {
      "post": {
        "tags": [
          "crate::controller::operator_controller"
        ],
        "operationId": "toggle_operator_status",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleOperatorStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "有効化状態変更完了"
          }
        },
        "security": [
          {
            "bearer_auth": [
              "Admin"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateListenerRequest": {
        "type": "object",
        "required": [
          "name",
          "lhost",
          "lport",
          "protocol"
        ],
        "properties": {
          "lhost": {
            "type": "string"
          },
          "lport": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "protocol": {
            "$ref": "#/components/schemas/ListenerType"
          }
        }
      },
      "GetOperatorRequest": {
        "type": "object",
        "required": [
          "operator_id"
        ],
        "properties": {
          "operator_id": {
            "type": "string"
          }
        }
      },
      "ListListenerResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "addr",
          "protocol"
        ],
        "properties": {
          "addr": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          }
        }
      },
      "ListenerType": {
        "type": "string",
        "enum": [
          "TCP",
          "HTTP",
          "HTTPS"
        ]
      },
      "OperatorCredential": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "RemoveListenerRequest": {
        "type": "object",
        "required": [
          "listener_id"
        ],
        "properties": {
          "listener_id": {
            "type": "string"
          }
        }
      },
      "StartListenerRequest": {
        "type": "object",
        "required": [
          "listener_id"
        ],
        "properties": {
          "listener_id": {
            "type": "string"
          }
        }
      },
      "StopListenerRequest": {
        "type": "object",
        "required": [
          "listener_id"
        ],
        "properties": {
          "listener_id": {
            "type": "string"
          }
        }
      },
      "ToggleOperatorStatusRequest": {
        "type": "object",
        "required": [
          "operator_id"
        ],
        "properties": {
          "operator_id": {
            "type": "string"
          }
        }
      },
      "UploadFile": {
        "type": "object",
        "required": [
          "filename",
          "data"
        ],
        "properties": {
          "data": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}